Extracted PIDJoint class into seperate code and header file.
